# docs/api-spec.yaml
openapi: 3.0.0
info:
  title: 就活サポートサービス API
  description: 就活サポートサービスのRESTful API仕様書
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: ローカル開発サーバー
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase IDトークンを使用
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        firebase_uid:
            type: string
            example: abc123def456ghi789
        name:
          type: string
          example: 山田太郎
        email:
          type: string
          example: yamada@example.com
        icon:
          type: string
          example: https://example.com/icon.png
        basic_es:
          type: string
          example: 私の強みは...
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Company:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 株式会社サンプル
        web_url:
          type: string
          example: https://example.com
        description:
          type: string
          example: ITソリューション企業です
        img:
          type: string
          example: https://example.com/logo.png
        industry:
          type: string
          example: IT・ソフトウェア
        scrape_target_url:
          type: string
          example: https://example.com/jobs
        last_scrape_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobEvent:
      type: object
      properties:
        id:
          type: integer
          example: 1
        company_id:
          type: integer
          example: 1
        job_title:
          type: string
          example: ソフトウェアエンジニア
        job_type:
          type: string
          example: 正社員
        job_description:
          type: string
          example: Webアプリケーション開発
        start_date:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        event_url:
          type: string
          example: https://example.com/job/1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Interview:
      type: object
      properties:
        id:
          type: integer
          example: 1
        company_id:
          type: integer
          example: 1
        job_event_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        interview_at:
          type: string
          format: date-time
        stage:
          type: string
          example: 一次面接
        audio_note_path:
          type: string
          example: /uploads/audio/001.wav
        transcript:
          type: string
          example: こんにちは、よろしく...
        audio_summary:
          type: string
          example: 技術的な質問が中心でした
        text_note:
          type: string
          example: 緊張したが良い印象だった
        location:
          type: string
          example: 東京オフィス2F会議室
        meeting_url:
          type: string
          example: https://meet.google.com/abc-defg-hij
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CompanyES:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        company_id:
          type: integer
          example: 1
        content:
          type: string
          example: 志望動機は...
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Success:
      type: object
      properties:
        message:
          type: string
          example: 操作が成功しました
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: ユーザー作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: 不正な入力

    get:
      tags:
        - Users
      summary: ユーザー一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'

    put:
      tags:
        - Users
      summary: ユーザー更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
  /users/{id}:
    get:
      tags:
        - Users
      summary: ユーザー詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: ユーザーが見つかりません

    delete:
      tags:
        - Users
      summary: ユーザー削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  # 企業関連（実装済み）
  /companies:
    post:
      tags:
        - Companies
      summary: 企業作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        '201':
          description: 作成成功

    get:
      tags:
        - Companies
      summary: 企業一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Company'

    put:
      tags:
        - Companies
      summary: 企業更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Company'
      responses:
        '200':
          description: 更新成功
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: 企業詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'

    delete:
      tags:
        - Companies
      summary: 企業削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功

  # 求人イベント関連（実装済み）
  /job-events:
    post:
      tags:
        - JobEvents
      summary: 求人イベント作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobEvent'
      responses:
        '201':
          description: 作成成功

    get:
      tags:
        - JobEvents
      summary: 求人イベント一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobEvent'

    put:
      tags:
        - JobEvents
      summary: 求人イベント更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobEvent'
      responses:
        '200':
          description: 更新成功
  /job-events/{id}:
    get:
      tags:
        - JobEvents
      summary: 求人イベント詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEvent'

    delete:
      tags:
        - JobEvents
      summary: 求人イベント削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
  /job-events/company/{companyID}:
    get:
      tags:
        - JobEvents
      summary: 企業別求人イベント取得
      parameters:
        - name: companyID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/JobEvent'

  # 面接関連（実装済み）
  /interviews:
    post:
      tags:
        - Interviews
      summary: 面接記録作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Interview'
      responses:
        '201':
          description: 作成成功

    get:
      tags:
        - Interviews
      summary: 面接記録一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Interview'

    put:
      tags:
        - Interviews
      summary: 面接記録更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Interview'
      responses:
        '200':
          description: 更新成功
  /interviews/{id}:
    get:
      tags:
        - Interviews
      summary: 面接記録詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interview'

    delete:
      tags:
        - Interviews
      summary: 面接記録削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
...
  /company-es:
    post:
      tags:
        - CompanyES
      summary: 企業ES作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyES'
      responses:
        '201':
          description: 作成成功

    get:
      tags:
        - CompanyES
      summary: 企業ES一覧取得
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyES'

    put:
      tags:
        - CompanyES
      summary: 企業ES更新
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyES'
      responses:
        '200':
          description: 更新成功
  /company-es/{id}:
    get:
      tags:
        - CompanyES
      summary: 企業ES詳細取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyES'

    delete:
      tags:
        - CompanyES
      summary: 企業ES削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 削除成功
  /company-es/user/{userID}:
    get:
      tags:
        - CompanyES
      summary: ユーザー別ES取得
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyES'
  /company-es/company/{companyID}:
    get:
      tags:
        - CompanyES
      summary: 企業別ES取得
      parameters:
        - name: companyID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyES'
  /company-es/user/{userID}/company/{companyID}:
    get:
      tags:
        - CompanyES
      summary: ユーザー・企業指定ES取得
      parameters:
        - name: userID
          in: path
          required: true
          schema:
            type: integer
        - name: companyID
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyES'

  # =========================
  # 以下、未実装のエンドポイント
  # =========================

  # 認証関連（未実装）
...